Внес исправления в домашню работу 2: Задание 3 - убрал не нужные кавычки; Задание 4 - переименовал локальную переменную в 'maxValue'; Ответ на вопрос: "что будет если передать new int[]{-1, -20}" будет 0. В новой редакции я установил сразу int maxValue  = Integer.MIN_VALUE; Задание 5 - -убрал приведение к int;

+++ Lesson2/src/Homework5.java
@@ -12,7 +12,7 @@ public class Homework5 {
 
     public static double calculateHypotenuse(int a, int b) {
         // a^{2}+b^{2}=c^{2}
-        int hypotenuse = (int) Math.sqrt((int) Math.pow(a, 2) + (int) Math.pow(b, 2));
+        double hypotenuse = Math.sqrt(Math.pow(a, 2) + Math.pow(b, 2));
         return hypotenuse;
     }
 }
